Revert "Merge remote-tracking branch 'origin/groupe_video' into groupe_video"
This reverts commit 677f16883ef90e2bf8d9d0725e8c280f4552c723.

+++ app/src/main/java/com/master/info_ua/videoannottool/annotation/Annotation.java
@@ -44,7 +44,7 @@ public class Annotation implements Serializable {
         this.annotationType = annotationType;
     }
 
-    public Annotation(String annotationTitle, Long annotationStartTime, Integer annotationDuration, AnnotationType annotationType, String audioFileName, String drawFileName, String textComment) {
+    public Annotation(String annotationTitle, Integer annotationStartTime, Integer annotationDuration, AnnotationType annotationType, String audioFileName, String drawFileName, String textComment) {
         this.annotationTitle = annotationTitle;
         this.annotationDate = new Date();
         this.annotationStartTime = annotationStartTime;
